REF: collect DTI/TDI/PI tests
