+++ pandas/tests/scalar/test_nat.py
@@ -546,3 +546,9 @@ def test_nat_addsub_tdlike_scalar(obj):
     assert NaT + obj is NaT
     assert obj + NaT is NaT
     assert NaT - obj is NaT
+
+
+def test_pickle():
+    # GH#4606
+    p = tm.round_trip_pickle(NaT)
+    assert p is NaT
